Refactor code and cleanup
Chatbot class have some behaviours that should be abstracted to the Ui
class instead.

The behaviours are not relevant to the Chatbot's purpose.

Extracting the behaviour into a distinct class for Ui allows
us to read and understand the code written better.

Let's pull behaviours not relevant to Chatbot out of it and move
it to another class.